third_party/WebKit/LayoutTests/media/controls/controls-video-keynav-space-vol.html

<DOCTYPE html>
<title>Test media controls video keyboard navigation for volume and space
  bar</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<video controls></video>
<script>
test(_ => {
  assert_true('eventSender' in window);
}, 'eventSender required');

async_test(function(t) {
  var video = document.querySelector("video");
  video.src = findMediaFile("video", "../content/test");
  assert_equals(video.volume, 1);
  assert_equals(video.currentTime, 0);

  // Focus the video.
  video.focus();

  video.oncanplaythrough = t.step_func_done(function() {
    video.onplay = t.step_func(() => {
      assert_false(video.paused);
      video.onvolumechange = t.step_func(() => {
        assert_equals(video.volume, .95);
      });
      // 'Down' reduces volume by .05.
      eventSender.keyDown("ArrowDown");

      video.onpause = t.step_func(() => {
        assert_true(video.paused);
        video.onvolumechange = t.step_func(() => {
          assert_equals(video.volume, 1);
        });
        // 'Up' increases volume by .05
        eventSender.keyDown("ArrowUp");
      });
      // 'Space' again pauses the video.
      eventSender.keyDown(" ");
    });
  });
  // 'Space' plays the video.
  eventSender.keyDown(" ");
});
</script>
